requirements.txt
click
semver
gitpython
pyyaml